chore: Refactoring code by taking out the common logic for re-using it on other IDEs (#39577)
