+++ app/client/src/IDE/utils/getNextEntityAfterRemove.test.ts
@@ -1,7 +1,10 @@
 import { EditorState } from "IDE/enums";
 import { PluginType } from "entities/Plugin";
 import * as FocusEntityObj from "navigation/FocusEntity";
-import { RedirectAction, getNextEntityAfterRemove } from "./IDESaga";
+import {
+  RedirectAction,
+  getNextEntityAfterRemove,
+} from "./getNextEntityAfterRemove";
 import { FocusEntity } from "navigation/FocusEntity";
 import type { EntityItem } from "ee/IDE/Interfaces/EntityItem";
